Make the sequence and server-time conversions explicit
Six more -Wsign-compare warnings are a uint32 measured against an int
that cannot hold a negative value, so the conversion C already performs
is the intended one and writing it out changes nothing.

Netchan_Process masks bit 31 off both wire values before using them, and
incoming_sequence, incoming_acknowledged and last_reliable_sequence are
only ever assigned 0 or one of those masked values - so all three stay
inside 0..2^31-1 despite being int. sv.time is a uint32 and svs.realtime
an int counting milliseconds up from zero.

Proof that this is a no-op: the GCC Release .text is byte-identical to a
control built from the previous commit, 267,636 bytes, after stripping
the build id.

Seven sign-compare warnings remain, all of them cases where the signed
side really can go negative.

+++ qcommon/net_chan.c
@@ -428,7 +428,7 @@ qboolean Netchan_Process (netchan_t *chan, sizebuf_t *msg)
 // discard stale or duplicated packets
 //
 
-	if (sequence <= chan->incoming_sequence)
+	if (sequence <= (uint32)chan->incoming_sequence)
 	{
 		if (showdrop->intvalue)
 			Com_Printf ("%s:Out of order packet %i at %i\n", LOG_NET
@@ -457,14 +457,14 @@ qboolean Netchan_Process (netchan_t *chan, sizebuf_t *msg)
 	chan->in_dropped += chan->dropped;
 
 	// MH: use acknowledgements to measure server->client packet loss (replaces sv_lag_stats option)
-	if (sequence_ack > chan->incoming_acknowledged)
+	if (sequence_ack > (uint32)chan->incoming_acknowledged)
 	{
-		if (chan->countacks && (curtime - chan->last_received < 90 || sequence_ack == chan->incoming_acknowledged + 1))
+		if (chan->countacks && (curtime - chan->last_received < 90 || sequence_ack == (uint32)(chan->incoming_acknowledged + 1)))
 		{
 			chan->out_total += sequence_ack - chan->incoming_acknowledged;
 			chan->out_dropped += sequence_ack - chan->incoming_acknowledged - 1;
 		}
-		else if (chan->reliable_length && sequence_ack >= chan->last_reliable_sequence)
+		else if (chan->reliable_length && sequence_ack >= (uint32)chan->last_reliable_sequence)
 		{
 			chan->out_total++;
 			if (reliable_ack != chan->reliable_sequence)
